fix: Webtoon source code has been refactored (around 80% of the code is refactored and made it smaller)
From 500 lines to 220 lines of code. (crazy refactoring)

The other changes are just some util function called method changed. Not a big deal

+++ cmd/metadata_gen/main.go
@@ -214,7 +214,7 @@ func parseChapterNumber(name string) (float32, bool) {
 		return -69, false
 	}
 	
-	return float32(f), true
+	return f, true
 }
 
 // eg: img_14.webp = 14
