+++ internal/sources/scanlators/madarascans.go
@@ -178,7 +178,7 @@ func (m *MadaraScans) extractChapterNumber(text string) float32 {
 
     match := result.FindStringSubmatch(text)
     if len(match) > 1 {
-        return float32(util.StringToFloat(strings.TrimSpace(match[1])))
+        return util.StringToFloat(strings.TrimSpace(match[1]))
     }
 
 	return 0
